fix: numpy 2.x 兼容 - np.in1d -> np.isin, np.int_ -> int
numpy 2.0 移除了 np.in1d (替代 np.isin) 和 np.int_ 别名。
stdweb 简单暂现源检测在 numpy 2.5 下报 AttributeError:
module 'numpy' has no attribute 'in1d'

+++ stdpipe/photometry.py
@@ -1082,7 +1082,7 @@ def get_objects_sep(
         # pixel outside each source's footprint, so the winpos aperture always
         # overlaps masked pixels by construction. Keep TRUNC / ALLMASKED /
         # NONPOSITIVE — those still convey real centroiding-quality info.
-        flag_win &= ~np.int_(sep.APER_HASMASKED)
+        flag_win &= ~int(sep.APER_HASMASKED)
     else:
         # Use initial positions from extraction
         xwin, ywin = obj0['x'], obj0['y']
